% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeTilegram2.R
\name{makeTilegram2}
\alias{makeTilegram2}
\title{Create a hexagonal map}
\usage{
makeTilegram2(sp, cellsize = NULL)
}
\arguments{
\item{cellsize}{The optional argument to change cellsize, Default is null, and probably best to leave empty}

\item{shp}{The spatial object to be read in, providing the coordinates and the dbf}
}
\value{
The transformed spatial object and data frame, though polygons now in hexagon format.
}
\description{
This function takes regular shapefiles with a population field and firt converts the maps to a hexagonal grid, with data preserved. Note that 
this function is heavily based off of the original, though is an improvement.
}
\examples{
mi_shp <- readOGR(eguia_path, "mi_county_shp")
test_hexagon <- hexcarto2b(mi_shp)

}
